+++ app/src/main/java/com/example/drivingefficiencyapp/trip/Trip.kt
@@ -1,5 +1,9 @@
 package com.example.drivingefficiencyapp.trip
 
+import com.google.firebase.firestore.DocumentId
+import com.google.firebase.firestore.ServerTimestamp
+import java.util.Date
+
 /**
  * Data class representing a trip with a date and duration.
  * Includes a companion object to store all trips in memory.
@@ -9,15 +13,10 @@ package com.example.drivingefficiencyapp.trip
  */
 
 data class Trip(
-    val date: String,
-    val duration: String
-) {
-    //Companion object to store all trips in memory, can be accessed statically from Trip class
-    companion object {
-        /**
-         * Static list to store all trips in memory.
-         * temporary solution as trips will be lost when the app closes.
-         */
-        val tripsList = mutableListOf<Trip>()
-    }
-}
+    @DocumentId
+    val id: String = "",
+    val date: String = "",
+    val duration: String = "",
+    @ServerTimestamp
+    val timestamp: Date? = null
+)